ui_service/requirements.txt
pandas
numpy
streamlit
requests
watchdog
google-cloud-storage